docs(app): add documentation for the playgrounds
