--- apps/ng-doc/src/app/docs/content/guidelines/playground/index.md
@@ -1,7 +1,0 @@
-# {{ NgDocPage.title }}
-
-It's time to write some awesome docs!
-
-## Playground
-
-{{ NgDocActions.playground("TagPlayground") }}
